Changing DataflowFramework To use Template
We are making DataflowFramework a template class so we can pass in the
type it's working on. Cpp templates can only have definitions in header
file, so we might delete dataflow.cpp if required later on, for now we
leave it and put the entire definition of DataflowFramework template in dataflow.h

In addition, we also added the expression vector to arguments we get in
the constructor. This is done because we want ownership or instantiation
of the domain set vector to be done externally. We should try to have
memory management (instantiation and destruction) of these in a common
place (usually a factory, but we don't need that in this case).

Depending on implementation details, we may pass in an already populated
set, or we can pass in the memory and then populate the set internally
in the DataflowFramework class.