Issue #26: Added Instruction level granularity to dataflow framework
This is required for faint analysis, because we need a unified form of
the transfer function to run on each line, from end of bb to beginning
of bb. This is so that depkill can make use of constkill and constgen
effects. Added additional flag and enum which uses BASIC_BLOCK
granularity or INSTRUCTION granularity, where BASIC_BLOCK is a default.

As a side effect of this, we added two overloaded version of genEval and
killEval which both take in Instruction* instead of BasicBlock*. There
should be no change required by existing passes.